src/third_party/blink/web_tests/external/wpt/css/css-pseudo/idlharness-expected.txt

This is a testharness.js-based test.
FAIL idl_test setup promise_test: Unhandled rejection with value: object "TypeError: window.getPseudoElements is not a function"
PASS Partial interface Window: original interface defined
FAIL CSSPseudoElement interface: existence and properties of interface object assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface object length assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface object name assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface: attribute type assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement interface: attribute style assert_own_property: self does not have own property "CSSPseudoElement" expected property "CSSPseudoElement" missing
FAIL CSSPseudoElement must be primary interface of beforeElements.item(0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL Stringification of beforeElements.item(0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElement interface: beforeElements.item(0) must inherit property "type" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElement interface: beforeElements.item(0) must inherit property "style" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: existence and properties of interface object assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface object length assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface object name assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: attribute length assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: operation item(unsigned long) assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList interface: operation getByType(CSSOMString) assert_own_property: self does not have own property "CSSPseudoElementList" expected property "CSSPseudoElementList" missing
FAIL CSSPseudoElementList must be primary interface of beforeElements assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL Stringification of beforeElements assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: beforeElements must inherit property "length" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: beforeElements must inherit property "item(unsigned long)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: calling item(unsigned long) on beforeElements with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: beforeElements must inherit property "getByType(CSSOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL CSSPseudoElementList interface: calling getByType(CSSOMString) on beforeElements with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: beforeElements is not defined"
FAIL Window interface: operation getPseudoElements(Element, CSSOMString) assert_own_property: global object missing non-static operation expected property "getPseudoElements" missing
FAIL Window interface: window must inherit property "getPseudoElements(Element, CSSOMString)" with the proper type assert_own_property: expected property "getPseudoElements" missing
FAIL Window interface: calling getPseudoElements(Element, CSSOMString) on window with too few arguments must throw TypeError assert_own_property: expected property "getPseudoElements" missing
Harness: the test ran to completion.
